fix float issue for postgres in position sql

+++ core/src/opds/v2_0/utils.rs
@@ -57,10 +57,10 @@ impl OPDSV2QueryExt for DatabaseConnection {
 			r"
 			WITH ranked AS (
 				SELECT m.id,
-				       COALESCE(
-				           CAST(mm.number AS REAL),
-				           CAST(RANK() OVER (ORDER BY m.name ASC) AS REAL)
-				       ) AS position
+				   COALESCE(
+				       CAST(mm.number AS DOUBLE PRECISION),
+				       CAST(RANK() OVER (ORDER BY m.name ASC) AS DOUBLE PRECISION)
+				   ) AS position
 				FROM media m
 				LEFT JOIN media_metadata mm ON mm.media_id = m.id
 				WHERE m.series_id = $1
